gcc  + all proto in .h

+++ ft_nb_len_base2.c
@@ -6,7 +6,7 @@
 /*   By: thifranc <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/03/12 16:10:29 by thifranc          #+#    #+#             */
-/*   Updated: 2016/03/31 10:30:13 by thifranc         ###   ########.fr       */
+/*   Updated: 2016/04/06 10:05:30 by thifranc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_nb_len_base2(unsigned long long int n, int base)
 {
-	if (n < base)
+	if (n < (unsigned long long int)base)
 		return (1);
 	else
 		return (1 + ft_nb_len_base2(n / base, base));
